+++ NEWS.md
@@ -4,7 +4,8 @@
 ## v0.6.0
 
 - `TypstError` renamed to `TypstCommandError`
-- `set_preamble` no longer returns its value
+- `TypstContext`
+- `preamble` and `set_preamble` replaced by `context::TypstContext` and `set_context`
 - New `TypstString(::TypstContext, ::Any)` constructor
 - `repr(::MIME"text/typst", ::TypstString)` now returns a `String` instead of a `TypstString`
 - `render` takes a `context = TypstContext()` instead of extra keyword parameters